refactored the folders to move controllers to HiraganaController instead of App
