+++ hiraganagenerator/src/main/java/org/generator/api/HiraganaController.java
@@ -1,0 +1,36 @@
+package org.generator.api;
+
+import org.generator.HiraganaGenerator;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.types.Hiragana;
+
+import java.util.HashMap;
+import java.util.Map;
+@RestController
+public class HiraganaController {
+        private final HiraganaGenerator hiraganaGenerator;
+        public HiraganaController(HiraganaGenerator hiraganaGenerator) {
+            this.hiraganaGenerator = hiraganaGenerator;
+        }
+        HashMap<String, String> hiraganaMap = new HashMap<>();
+        @GetMapping("/generateHiragana")
+        public HashMap<String,String> generateHiragana(@RequestParam("size") int size) {
+            hiraganaMap.clear();
+            return getHiragana(size);
+        }
+
+        private HashMap<String, String> getHiragana(int size) {;
+
+            var hiraganaPair = hiraganaGenerator.hiraganaGenerator(size);
+
+            for (Map.Entry<Hiragana,String> entry : hiraganaPair.entrySet()) {
+                String englishValue = entry.getKey().getEnglishValue();
+                String hiraganaValue = entry.getKey().getValue();
+                hiraganaMap.put(hiraganaValue, englishValue);
+            }
+
+            return hiraganaMap;
+        }
+    }
