+++ kana-crazy/src/main/java/org/example/App.java
@@ -42,30 +42,3 @@ public class App {
         SpringApplication.run(App.class, args);
     }
 }
-
-@RestController
-class HiraganaController {
-    private final HiraganaGenerator hiraganaGenerator;
-    public HiraganaController(HiraganaGenerator hiraganaGenerator) {
-        this.hiraganaGenerator = hiraganaGenerator;
-    }
-    HashMap<String, String> hiraganaMap = new HashMap<>();
-    @GetMapping("/generateHiragana")
-    public HashMap<String,String> generateHiragana(@RequestParam("size") int size) {
-        hiraganaMap.clear();
-        return getHiragana(size);
-    }
-
-    private HashMap<String, String> getHiragana(int size) {;
-
-        var hiraganaPair = hiraganaGenerator.hiraganaGenerator(size);
-
-        for (Map.Entry<Hiragana,String> entry : hiraganaPair.entrySet()) {
-            String englishValue = entry.getKey().getEnglishValue();
-            String hiraganaValue = entry.getKey().getValue();
-            hiraganaMap.put(hiraganaValue, englishValue);
-        }
-
-        return hiraganaMap;
-    }
-}
